requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
praw==7.7.1
groq==0.4.2
python-dotenv==1.0.0
werkzeug==3.0.1
pandas==2.1.4